requirements.txt
flask==2.3.3
markdown==3.4.4
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
argparse==1.4.0
